Have it return the replicated movement speed
Are you not sending the WalkSpeed to the server when it's in the air state? If you are, then this shouldn't cause any errors and would be preferred instead of a flat float value. Let me know if that's okay or what needs to be implemented to fix this, I wanna implement this for bhopping!